更新了全新的credit数据层
1.修改了Po和表结构
2.删除了DaoImpl的update方法 改为add
即只添加数据不再删除数据 （因为要有记录）
3.提供set方法初始化新注册客户信用值
4.getcredit返回为list 返回为该客户所有的信用记录
单一返回最新的creditPo需要逻辑实现（注释有说明）

+++ OBSH_/src/data/datahelper/impl/UserDataMysqlHelper.java
@@ -36,7 +36,7 @@ public class UserDataMysqlHelper implements UserDataHelper {
 	            String name = ret.getString(2);  
 	            String password = ret.getString(3);  
 	            String contactinfo = ret.getString(4); 
-	            int credit = ret.getInt(5);
+	            double credit = ret.getInt(5);
 	            boolean isVIP = false;
 	            if(ret.getInt(6)==1)
 	            	isVIP = true;
